Move Link header emission from middleware to ResponseEmitter
Per review feedback, Link headers are now emitted by ResponseEmitter
(similar to cookie handling) instead of requiring a separate middleware.
Removed LinkHeaderMiddleware and added tests to ResponseEmitterTest.
